HHT-1174: moving common files
