+++ app/core/kg/kg_repository.py
@@ -11,7 +11,7 @@ from app.core.types import KGSliceId
 from app.kg.graph import Graph
 from app.serialize.jsonld_configuration import JsonLDConfiguration
 from app.serialize.jsonld_serializer import JsonLDSerialializer
-from app.transform.k8s.upper_ontology_base import UpperOntologyBase
+from app.transform.upper_ontology_base import UpperOntologyBase
 
 T = TypeVar("T")
