+++ app/core/test_graph_fixture.py
@@ -8,7 +8,7 @@ from app.kg.id_base import IdBase
 from app.kg.inmemory_graph import InMemoryGraph
 from app.kg.iri import IRI
 from app.serialize.jsonld_configuration import JsonLDConfiguration
-from app.transform.k8s.upper_ontology_base import UpperOntologyBase
+from app.transform.upper_ontology_base import UpperOntologyBase
 
 
 class TestTransformer(UpperOntologyBase):
