+++ app/transform/k8s/resource_termination_transformer.py
@@ -5,8 +5,8 @@ from app.core.kg.kg_snapshot import KGSnapshot
 from app.kg.graph import Graph
 from app.kg.iri import IRI
 from app.transform.k8s.transformation_context import TransformationContext
-from app.transform.k8s.transformer_base import TransformerBase
-from app.transform.k8s.upper_ontology_base import UpperOntologyBase
+from app.transform.transformer_base import TransformerBase
+from app.transform.upper_ontology_base import UpperOntologyBase
 
 
 class ResourceTerminationTransformer(UpperOntologyBase):
