+++ app/transform/k8s/workload_transformer.py
@@ -3,9 +3,9 @@ from typing import Any, Dict, List, Optional
 from app.kg.graph import Graph
 from app.kg.iri import IRI
 from app.transform.k8s.transformation_context import TransformationContext
-from app.transform.k8s.transformer_base import TransformerBase
-from app.transform.k8s.upper_ontology_base import UpperOntologyBase
 from app.transform.measurement import Measurement
+from app.transform.transformer_base import TransformerBase
+from app.transform.upper_ontology_base import UpperOntologyBase
 from app.util.quantity import parse_quantity
