+++ app/transform/metrics/node_metric_transformer.py
@@ -5,9 +5,9 @@ from app.core.types import MetricValue
 from app.kg.graph import Graph
 from app.kg.iri import IRI
 from app.transform.k8s.transformation_context import TransformationContext
-from app.transform.k8s.transformer_base import TransformerBase
-from app.transform.k8s.upper_ontology_base import Aggregation, UpperOntologyBase
 from app.transform.metrics.metric_transformer import MetricToGraphTransformerBase
+from app.transform.transformer_base import TransformerBase
+from app.transform.upper_ontology_base import Aggregation, UpperOntologyBase
 
 
 class NodeMetricToGraphTransformer(MetricToGraphTransformerBase, UpperOntologyBase):
